requirements.txt
azure-functions
azure-storage-queue
pymysql
azure-storage-blob